finished up the shift assignment card
